+++ src/compile_time.rs
@@ -1,127 +1,160 @@
-#[macro_export]
-macro_rules! flush {
-    () => {{
-        $crate::LOG_SENDER.try_send($crate::LogMessage::Flush).unwrap();
-    }};
-}
+pub const TRACE: usize = 5;
+pub const DEBUG: usize = 4;
+pub const INFO: usize = 3;
+pub const WARN: usize = 2;
+pub const ERROR: usize = 1;
+pub const OFF: usize = 0;
+
+pub const MAX_LEVEL: usize = if cfg!(feature = "max-level-trace") {
+    TRACE
+} else if cfg!(feature = "max-level-debug") {
+    DEBUG
+} else if cfg!(feature = "max-level-info") {
+    INFO
+} else if cfg!(feature = "max-level-warn") {
+    WARN
+} else if cfg!(feature = "max-level-error") {
+    ERROR
+} else if cfg!(feature = "max-level-off") {
+    OFF
+} else {
+    TRACE
+};
 
 #[macro_export]
-macro_rules! log_with_level {
+macro_rules! log_with_level_ct {
     // Case 1: topic, format string argument arguments, and key-value pairs
     ($level:expr, $topic:expr; $fmt:expr, $($arg:expr),* ; $($key:ident = $value:expr),+ $(,)?) => {{
-        $crate::log_fn_json_v2!($level, $topic; $fmt, $($arg),*; $($key = $value),+);
+        $crate::log_fn_json_v3!($level, $topic; $fmt, $($arg),*; $($key = $value),+);
     }};
 
     // Case 2: topic and static string, and key-value pairs
     ($level:expr, $topic:expr; $msg:expr; $($key:ident = $value:expr),+ $(,)?) => {{
-        $crate::log_fn_json_v2!($level, $topic; $msg; $($key = $value),+);
+        $crate::log_fn_json_v3!($level, $topic; $msg; $($key = $value),+);
     }};
 
     // Case 3: topic and key-value pairs
     ($level:expr, $topic:expr; $($key:ident = $value:expr),+ $(,)?) => {{
-        $crate::log_fn_json_v2!($level, $topic; ""; $($key = $value),+);
+        $crate::log_fn_json_v3!($level, $topic; ""; $($key = $value),+);
     }};
     
     // Case 4: topic and format string with arguments
     ($level:expr, $topic:expr; $fmt:expr, $($arg:expr),* $(,)?) => {{
-        $crate::log_fn_json_v2!($level, $topic; $fmt, $($arg),*);
+        $crate::log_fn_json_v3!($level, $topic; $fmt, $($arg),*);
     }};
 
     // Case 5: topic and static string
     ($level:expr, $topic:expr; $msg:expr) => {{
-        $crate::log_fn_json_v2!($level, $topic; $msg);
+        $crate::log_fn_json_v3!($level, $topic; $msg);
     }};
 
     // Case 6: Topic only
     ($level:expr, $topic:expr) => {{
-        $crate::log_fn_json_v2!($level, $topic; "");
+        $crate::log_fn_json_v3!($level, $topic; "");
     }};
 
     // **Case 7: Single key-value pair without topic**
     ($level:expr, $key:ident = $value:expr $(,)?) => {{
-        $crate::log_fn_json_v2!($level, $key = $value);
+        $crate::log_fn_json_v3!($level, $key = $value);
     }};
     
     // **Case 8: Multiple key-value pairs without topic**
     ($level:expr, $($key:ident = $value:expr),+ $(,)?) => {{
-        $crate::log_fn_json_v2!($level, $($key = $value),+);
+        $crate::log_fn_json_v3!($level, $($key = $value),+);
     }};
 }
 
 #[macro_export]
-macro_rules! flash_trace {
+macro_rules! flash_trace_ct {
     // Handle one or more key-value pairs without a topic
     ( $( $key:ident = $value:expr ),+ $(,)? ) => {
-        $crate::log_with_level!($crate::LogLevel::Trace, ""; $( $key = $value ),+ );
+        $crate::log_with_level_ct!($crate::compile_time::TRACE, ""; $( $key = $value ),+ );
     };
 
     // Handle all other cases (e.g., with topic, message, etc.)
     ( $($args:tt)* ) => {
-        $crate::log_with_level!($crate::LogLevel::Trace, $($args)* );
+        $crate::log_with_level_ct!($crate::compile_time::TRACE, $($args)* );
     };
 }
 
 #[macro_export]
-macro_rules! flash_debug {
+macro_rules! flash_debug_ct {
     // Handle one or more key-value pairs without a topic
     ( $( $key:ident = $value:expr ),+ $(,)? ) => {
-        $crate::log_with_level!($crate::LogLevel::Debug, ""; $( $key = $value ),+ )
+        $crate::log_with_level_ct!($crate::compile_time::DEBUG, ""; $( $key = $value ),+ )
     };
     // Handle all other cases (e.g., with topic, message, etc.)
     ( $($args:tt)* ) => {
-        $crate::log_with_level!($crate::LogLevel::Debug, $($args)* )
+        $crate::log_with_level_ct!($crate::compile_time::DEBUG, $($args)* )
     };
 }
 
 #[macro_export]
-macro_rules! flash_info {
+macro_rules! flash_info_ct {
     // Handle one or more key-value pairs without a topic
     ( $( $key:ident = $value:expr ),+ $(,)? ) => {
-        $crate::log_with_level!($crate::LogLevel::Info, ""; $( $key = $value ),+ );
+        $crate::log_with_level_ct!($crate::compile_time::INFO, ""; $( $key = $value ),+ );
     };
     // Handle all other cases (e.g., with topic, message, etc.)
     ( $($args:tt)* ) => {
-        $crate::log_with_level!($crate::LogLevel::Info, $($args)* );
+        $crate::log_with_level_ct!($crate::compile_time::INFO, $($args)* );
     };
 }
 
 #[macro_export]
-macro_rules! flash_warn {
+macro_rules! flash_warn_ct {
     // Handle one or more key-value pairs without a topic
     ( $( $key:ident = $value:expr ),+ $(,)? ) => {
-        $crate::log_with_level!($crate::LogLevel::Warn, ""; $( $key = $value ),+ )
+        $crate::log_with_level_ct!($crate::compile_time::WARN, ""; $( $key = $value ),+ )
     };
     // Handle all other cases (e.g., with topic, message, etc.)
     ( $($args:tt)* ) => {
-        $crate::log_with_level!($crate::LogLevel::Warn, $($args)* )
+        $crate::log_with_level_ct!($crate::compile_time::WARN, $($args)* )
     };
 }
 
 #[macro_export]
-macro_rules! flash_error {
+macro_rules! flash_error_ct {
     // Handle one or more key-value pairs without a topic
     ( $( $key:ident = $value:expr ),+ $(,)? ) => {
-        $crate::log_with_level!($crate::LogLevel::Error, ""; $( $key = $value ),+ )
+        $crate::log_with_level_ct!($crate::compile_time::ERROR, ""; $( $key = $value ),+ )
     };
     // Handle all other cases (e.g., with topic, message, etc.)
     ( $($args:tt)* ) => {
-        $crate::log_with_level!($crate::LogLevel::Error, $($args)* )
+        $crate::log_with_level_ct!($crate::compile_time::ERROR, $($args)* )
     };
 }
 
+#[inline]
+pub fn usize_to_level(level: usize) -> &'static str {
+    match level {
+        5 => "Trace",
+        4 => "Debug",
+        3 => "Info",
+        2 => "Warn",
+        1 => "Error",
+        0 => "Off",
+        _ => "Unknown",
+    }
+}
+
 #[macro_export]
-macro_rules! log_fn_json_v2 {
+macro_rules! log_fn_json_v3 {
     // Case 1: topic, format sring, kv
     ($level:expr, $topic:expr; $fmt:expr, $($arg:expr),*; $($key:ident = $value:expr),+ $(,)?) => {{
-        if $level <= $crate::LogLevel::from_usize($crate::MAX_LOG_LEVEL.load(std::sync::atomic::Ordering::Relaxed)).unwrap() {
+        if $level <= $crate::compile_time::MAX_LEVEL {
+            $(
+                #[allow(non_snake_case)]
+                let $key = $value.clone();
+            )+
             let func = move || {
-                let unixnano = $crate::get_unix_nano();
-                let include_unixnano = $crate::logger::INCLUDE_UNIXNANO.load(std::sync::atomic::Ordering::Relaxed);
                 let json_obj = $crate::serde_json::json!({
                     $(
-                        stringify!($key): $value,
+                        stringify!($key): $key,
                     )+
                 });
+                let unixnano = $crate::get_unix_nano();
+                let include_unixnano = $crate::logger::INCLUDE_UNIXNANO.load(std::sync::atomic::Ordering::Relaxed);
                 let timezone = $crate::TIMEZONE.load(std::sync::atomic::Ordering::Relaxed);
                 let (date, time) = $crate::convert_unix_nano_to_date_and_time(unixnano, timezone);
                 
@@ -130,7 +163,7 @@ macro_rules! log_fn_json_v2 {
                         "date": date,
                         "time": time,
                         "offset": timezone,
-                        "level": $level.to_string(),
+                        "level": $crate::compile_time::usize_to_level($level),
                         "src": format!("{}:{}", file!(), line!()),
                         "topic": $topic,
                         "data": json_obj,
@@ -140,7 +173,7 @@ macro_rules! log_fn_json_v2 {
                         "date": date,
                         "time": time,
                         "offset": timezone,
-                        "level": $level.to_string(),
+                        "level": $crate::compile_time::usize_to_level($level),
                         "src": format!("{}:{}", file!(), line!()),
                         "topic": $topic,
                         "data": json_obj,
@@ -158,15 +191,20 @@ macro_rules! log_fn_json_v2 {
     
     // Case 2: topic, static string, kv
     ($level:expr, $topic:expr; $msg:expr; $($key:ident = $value:expr),+ $(,)?) => {{
-        if $level <= $crate::LogLevel::from_usize($crate::MAX_LOG_LEVEL.load(std::sync::atomic::Ordering::Relaxed)).unwrap() {
+        if $level <= $crate::compile_time::MAX_LEVEL {
+            
+            $(
+                #[allow(non_snake_case)]
+                let $key = $value.clone();
+            )*
             let func = move || {
-                let unixnano = $crate::get_unix_nano();
-                let include_unixnano = $crate::logger::INCLUDE_UNIXNANO.load(std::sync::atomic::Ordering::Relaxed);
                 let json_obj = $crate::serde_json::json!({
                     $(
-                        stringify!($key): $value,
+                        stringify!($key): $key,
                     )+
                 });
+                let unixnano = $crate::get_unix_nano();
+                let include_unixnano = $crate::logger::INCLUDE_UNIXNANO.load(std::sync::atomic::Ordering::Relaxed);
                 let timezone = $crate::TIMEZONE.load(std::sync::atomic::Ordering::Relaxed);
                 let (date, time) = $crate::convert_unix_nano_to_date_and_time(unixnano, timezone);
                 let json_msg = match include_unixnano {
@@ -174,7 +212,7 @@ macro_rules! log_fn_json_v2 {
                         "date": date,
                         "time": time,
                         "offset": timezone,
-                        "level": $level.to_string(),
+                        "level": $crate::compile_time::usize_to_level($level),
                         "src": format!("{}:{}", file!(), line!()),
                         "topic": $topic,
                         "data": json_obj,
@@ -184,7 +222,7 @@ macro_rules! log_fn_json_v2 {
                         "date": date,
                         "time": time,
                         "offset": timezone,
-                        "level": $level.to_string(),
+                        "level": $crate::compile_time::usize_to_level($level),
                         "src": format!("{}:{}", file!(), line!()),
                         "topic": $topic,
                         "data": json_obj,
@@ -202,18 +240,18 @@ macro_rules! log_fn_json_v2 {
     
     // Case 3: topic and formated string
     ($level:expr, $topic:expr; $fmt:expr, $($arg:expr),* $(,)?) => {{
-        if $level <= $crate::LogLevel::from_usize($crate::MAX_LOG_LEVEL.load(std::sync::atomic::Ordering::Relaxed)).unwrap() {
+        if $level <= $crate::compile_time::MAX_LEVEL {
             let func = move || {
-                let unixnano = $crate::get_unix_nano();
-                let include_unixnano = $crate::logger::INCLUDE_UNIXNANO.load(std::sync::atomic::Ordering::Relaxed);
                 let timezone = $crate::TIMEZONE.load(std::sync::atomic::Ordering::Relaxed);
+                let unixnano = $crate::get_unix_nano();
                 let (date, time) = $crate::convert_unix_nano_to_date_and_time(unixnano, timezone);
+                let include_unixnano = $crate::logger::INCLUDE_UNIXNANO.load(std::sync::atomic::Ordering::Relaxed);
                 let json_msg = match include_unixnano {
                     false => $crate::serde_json::json!({
                         "date": date,
                         "time": time,
                         "offset": timezone,
-                        "level": $level.to_string(),
+                        "level": $crate::compile_time::usize_to_level($level),
                         "src": format!("{}:{}", file!(), line!()),
                         "topic": $topic,
                         "message": format!($fmt, $($arg),*),
@@ -223,7 +261,7 @@ macro_rules! log_fn_json_v2 {
                         "date": date,
                         "time": time,
                         "offset": timezone,
-                        "level": $level.to_string(),
+                        "level": $crate::compile_time::usize_to_level($level),
                         "src": format!("{}:{}", file!(), line!()),
                         "topic": $topic,
                         "message": format!($fmt, $($arg),*),
@@ -239,7 +277,7 @@ macro_rules! log_fn_json_v2 {
     
     // Case 4: topic and static string
     ($level:expr, $topic:expr; $msg:expr $(,)?) => {{
-        if $level <= $crate::LogLevel::from_usize($crate::MAX_LOG_LEVEL.load(std::sync::atomic::Ordering::Relaxed)).unwrap() {
+        if $level <= $crate::compile_time::MAX_LEVEL {
             let func = move || {
                 let unixnano = $crate::get_unix_nano();
                 let include_unixnano = $crate::logger::INCLUDE_UNIXNANO.load(std::sync::atomic::Ordering::Relaxed);
@@ -250,7 +288,7 @@ macro_rules! log_fn_json_v2 {
                         "date": date,
                         "time": time,
                         "offset": timezone,
-                        "level": $level.to_string(),
+                        "level": $crate::compile_time::usize_to_level($level),
                         "src": format!("{}:{}", file!(), line!()),
                         "topic": $topic.to_string(),
                         "message": $msg
@@ -259,7 +297,7 @@ macro_rules! log_fn_json_v2 {
                         "date": date,
                         "time": time,
                         "offset": timezone,
-                        "level": $level.to_string(),
+                        "level": $crate::compile_time::usize_to_level($level),
                         "src": format!("{}:{}", file!(), line!()),
                         "topic": $topic.to_string(),
                         "message": $msg,
@@ -276,13 +314,17 @@ macro_rules! log_fn_json_v2 {
 
     // **Case 7: Single key-value pair without topic**
     ($level:expr, $key:ident = $value:expr) => {{
-        if $level <= $crate::LogLevel::from_usize($crate::MAX_LOG_LEVEL.load(std::sync::atomic::Ordering::Relaxed)).unwrap() {
+        if $level <= $crate::compile_time::MAX_LEVEL {
+            $(
+                #[allow(non_snake_case)]
+                let $key = $value.clone();
+            )*
             let func = move || {
                 let unixnano = $crate::get_unix_nano();
                 let include_unixnano = $crate::logger::INCLUDE_UNIXNANO.load(std::sync::atomic::Ordering::Relaxed);
                 let json_obj = $crate::serde_json::json!({
                     $(
-                        stringify!($key): $value,
+                        stringify!($key): $key,
                     )+
                 });
                 let timezone = $crate::TIMEZONE.load(std::sync::atomic::Ordering::Relaxed);
@@ -292,7 +334,7 @@ macro_rules! log_fn_json_v2 {
                         "date": date,
                         "time": time,
                         "offset": timezone,
-                        "level": $level.to_string(),
+                        "level": $crate::compile_time::usize_to_level($level),
                         "src": format!("{}:{}", file!(), line!()),
                         "topic": "",
                         "data": json_obj,
@@ -302,7 +344,7 @@ macro_rules! log_fn_json_v2 {
                         "date": date,
                         "time": time,
                         "offset": timezone,
-                        "level": $level.to_string(),
+                        "level": $crate::compile_time::usize_to_level($level),
                         "src": format!("{}:{}", file!(), line!()),
                         "topic": "",
                         "data": json_obj,
@@ -319,13 +361,17 @@ macro_rules! log_fn_json_v2 {
     
     // **Case 8: Multiple key-value pairs without topic**
     ($level:expr, $($key:ident = $value:expr),+ $(,)?) => {{
-        if $level <= $crate::LogLevel::from_usize($crate::MAX_LOG_LEVEL.load(std::sync::atomic::Ordering::Relaxed)).unwrap() {
+        if $level <= $crate::compile_time::MAX_LEVEL {
+            $(
+                #[allow(non_snake_case)]
+                let $key = $value.clone();
+            )*
             let func = move || {
                 let unixnano = $crate::get_unix_nano();
                 let include_unixnano = $crate::logger::INCLUDE_UNIXNANO.load(std::sync::atomic::Ordering::Relaxed);
                 let json_obj = $crate::serde_json::json!({
                     $(
-                        stringify!($key): $value,
+                        stringify!($key): $key,
                     )+
                 });
                 let timezone = $crate::TIMEZONE.load(std::sync::atomic::Ordering::Relaxed);
@@ -335,7 +381,7 @@ macro_rules! log_fn_json_v2 {
                         "date": date,
                         "time": time,
                         "offset": timezone,
-                        "level": $level.to_string(),
+                        "level": $crate::compile_time::usize_to_level($level),
                         "src": format!("{}:{}", file!(), line!()),
                         "data": json_obj,
                         "topic": "",
@@ -345,7 +391,7 @@ macro_rules! log_fn_json_v2 {
                         "date": date,
                         "time": time,
                         "offset": timezone,
-                        "level": $level.to_string(),
+                        "level": $crate::compile_time::usize_to_level($level),
                         "src": format!("{}:{}", file!(), line!()),
                         "data": json_obj,
                         "topic": "",
@@ -363,38 +409,39 @@ macro_rules! log_fn_json_v2 {
 
 #[cfg(test)]
 mod tests {
+    use crate::logger::Logger;
+    use crate::logger::LogLevel;
+    use crate::logger::TimeZone;
     use anyhow::Result;
     use serde::Serialize;
-    use crate::{Logger, LogLevel};
-    use crate::TimeZone;
 
     #[derive(Debug, Clone, Serialize)]
     pub enum Hello {
-        World,
         FlashLog,
+        World,
     }
 
     impl std::fmt::Display for Hello {
         fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
             match self {
-                Hello::World => write!(f, "World"),
                 Hello::FlashLog => write!(f, "FlashLog"),
+                Hello::World => write!(f, "World"),
             }
         }
     }
 
-    #[derive(Debug, Serialize, Clone)]
+    #[derive(Debug, Clone, Serialize)]
     pub struct TestStruct {
-        pub test: i32,
+        test: i32,
     }
 
-    #[derive(Debug, Serialize, Clone)]
+    #[derive(Debug, Clone, Serialize)]
     pub struct TestStruct2 {
-        pub test: i32,
+        test: i32,
     }
 
     #[test]
-    fn test_logger() -> Result<()> {
+    fn test_ct() -> Result<()> {
         let _guard = Logger::initialize()
             //.with_file("logs", "message")?
             .with_console_report(true)
@@ -405,80 +452,85 @@ mod tests {
             .include_unixnano(false)
             .launch();
 
-        flash_error!(Hello::FlashLog);
-        flash_error!(Hello::World);
-        flash_error!("Hello");
-        flash_error!("Hello"; "FlashLog");
-        flash_error!("Hello"; "FlashLog"; version = "0.1.0");
-        flash_error!("Hello"; "FlashLog"; version = "0.1.0", author = "John Doe");
-        flash_error!(version = "0.1.0");
-        flash_error!(version = "0.1.0", author = "John Doe");
-        flash_error!("topic1"; "message {} {}", 1, 2);
-        flash_error!("topic2"; "message {} {}", 1, 2; struct_info = 1, struct_info2 = 2);
-        flash_error!("topic2"; "message {} {}", 1, 2; Hello = Hello::FlashLog);
+        flash_error_ct!(Hello::FlashLog);
+        flash_error_ct!(Hello::World);
+        flash_error_ct!("Hello");
+        flash_error_ct!("Hello"; "FlashLog");
+        flash_error_ct!("Hello"; "FlashLog"; version = "0.1.0");
+        flash_error_ct!("Hello"; "FlashLog"; version = "0.1.0", author = "John Doe");
+        flash_error_ct!(version = "0.1.0");
+        flash_error_ct!(version = "0.1.0", author = "John Doe");
+        flash_error_ct!("topic1"; "message {} {}", 1, 2);
+        flash_error_ct!("topic2"; "message {} {}", 1, 2; struct_info = 1, struct_info2 = 2);
+        flash_error_ct!("topic2"; "message {} {}", 1, 2; Hello = Hello::FlashLog);
         let test_info = TestStruct { test: 1 };
         let test_info2 = TestStruct2 { test: 2 };
-        flash_error!("topic2"; "message {} {}", 1, 2; TestStruct = test_info, TestStruct2 = test_info2);
-        flush!(); // this flushes regardless of the buffer size and flush interval
+        flash_error_ct!("topic2"; "message {} {}", 1, 2; TestStruct = test_info, TestStruct2 = test_info2);
+        println!("{:?}", test_info); // still alive
+        crate::flush!(); // this flushes regardless of the buffer size and flush interval
 
-        flash_warn!(Hello::World);
-        flash_warn!("Hello");
-        flash_warn!("Hello"; "FlashLog");
-        flash_warn!("Hello"; "FlashLog"; version = "0.1.0");
-        flash_warn!("Hello"; "FlashLog"; version = "0.1.0", author = "John Doe");
-        flash_warn!(version = "0.1.0");
-        flash_warn!(version = "0.1.0", author = "John Doe");
-        flash_warn!("topic1"; "message {} {}", 1, 2);
-        flash_warn!("topic2"; "message {} {}", 1, 2; struct_info = 1, struct_info2 = 2);
-        flash_warn!("topic2"; "message {} {}", 1, 2; Hello = Hello::FlashLog);
+        flash_warn_ct!(Hello::World);
+        flash_warn_ct!("Hello");
+        flash_warn_ct!("Hello"; "FlashLog");
+        flash_warn_ct!("Hello"; "FlashLog"; version = "0.1.0");
+        flash_warn_ct!("Hello"; "FlashLog"; version = "0.1.0", author = "John Doe");
+        flash_warn_ct!(version = "0.1.0");
+        flash_warn_ct!(version = "0.1.0", author = "John Doe");
+        flash_warn_ct!("topic1"; "message {} {}", 1, 2);
+        flash_warn_ct!("topic2"; "message {} {}", 1, 2; struct_info = 1, struct_info2 = 2);
+        flash_warn_ct!("topic2"; "message {} {}", 1, 2; Hello = Hello::FlashLog);
         let test_info = TestStruct { test: 1 };
-        flash_warn!("topic2"; "message {} {}", 1, 2; TestStruct = test_info);
+        flash_warn_ct!("topic2"; "message {} {}", 1, 2; TestStruct = test_info);
+        println!("{:?}", test_info);
 
-        flush!(); // this flushes regardless of the buffer size and flush interval
+        crate::flush!(); // this flushes regardless of the buffer size and flush interval
 
-        flash_info!(Hello::World);
-        flash_info!(Hello::FlashLog);
-        flash_info!("Hello"); 
-        flash_info!("Hello"; "FlashLog");
-        flash_info!("Hello"; "FlashLog"; version = "0.1.0");
-        flash_info!("Hello"; "FlashLog"; version = "0.1.0", author = "John Doe");
-        flash_info!(version = "0.1.0");
-        flash_info!(version = "0.1.0", author = "John Doe");
-        flash_info!("topic1"; "message {} {}", 1, 2);
-        flash_info!("topic2"; "message {} {}", 1, 2; struct_info = 1, struct_info2 = 2);
-        flash_info!("topic2"; "message {} {}", 1, 2; Hello = Hello::FlashLog);
+        flash_info_ct!(Hello::World);
+        flash_info_ct!(Hello::FlashLog);
+        flash_info_ct!("Hello"); 
+        flash_info_ct!("Hello"; "FlashLog");
+        flash_info_ct!("Hello"; "FlashLog"; version = "0.1.0");
+        flash_info_ct!("Hello"; "FlashLog"; version = "0.1.0", author = "John Doe");
+        flash_info_ct!(version = "0.1.0");
+        flash_info_ct!(version = "0.1.0", author = "John Doe");
+        flash_info_ct!("topic1"; "message {} {}", 1, 2);
+        flash_info_ct!("topic2"; "message {} {}", 1, 2; struct_info = 1, struct_info2 = 2);
+        flash_info_ct!("topic2"; "message {} {}", 1, 2; Hello = Hello::FlashLog);
         let test_info = TestStruct { test: 1 };
-        flash_info!("topic2"; "message {} {}", 1, 2; TestStruct = test_info);
+        flash_info_ct!("topic2"; "message {} {}", 1, 2; TestStruct = test_info);
+        println!("{:?}", test_info);
 
-        flush!(); // this flushes regardless of the buffer size and flush interval
+        crate::flush!(); // this flushes regardless of the buffer size and flush interval
 
-        flash_debug!(Hello::World);
-        flash_debug!("Hello");
-        flash_debug!("Hello"; "FlashLog");
-        flash_debug!("Hello"; "FlashLog"; version = "0.1.0");
-        flash_debug!("Hello"; "FlashLog"; version = "0.1.0", author = "John Doe");
-        flash_debug!(version = "0.1.0");
-        flash_debug!(version = "0.1.0", author = "John Doe");
-        flash_debug!("topic1"; "message {} {}", 1, 2);
-        flash_debug!("topic2"; "message {} {}", 1, 2; struct_info = 1, struct_info2 = 2);
-        flash_debug!("topic2"; "message {} {}", 1, 2; Hello = Hello::FlashLog);
+        flash_debug_ct!(Hello::World);
+        flash_debug_ct!("Hello");
+        flash_debug_ct!("Hello"; "FlashLog");
+        flash_debug_ct!("Hello"; "FlashLog"; version = "0.1.0");
+        flash_debug_ct!("Hello"; "FlashLog"; version = "0.1.0", author = "John Doe");
+        flash_debug_ct!(version = "0.1.0");
+        flash_debug_ct!(version = "0.1.0", author = "John Doe");
+        flash_debug_ct!("topic1"; "message {} {}", 1, 2);
+        flash_debug_ct!("topic2"; "message {} {}", 1, 2; struct_info = 1, struct_info2 = 2);
+        flash_debug_ct!("topic2"; "message {} {}", 1, 2; Hello = Hello::FlashLog);
         let test_info = TestStruct { test: 1 };
-        flash_debug!("topic2"; "message {} {}", 1, 2; TestStruct = test_info);
+        flash_debug_ct!("topic2"; "message {} {}", 1, 2; TestStruct = test_info);
+        println!("{:?}", test_info);
 
-        flush!(); // this flushes regardless of the buffer size and flush interval
+        crate::flush!(); // this flushes regardless of the buffer size and flush interval
 
-        flash_trace!(Hello::World);
-        flash_trace!("Hello");
-        flash_trace!("Hello"; "FlashLog");
-        flash_trace!("Hello"; "FlashLog"; version = "0.1.0");
-        flash_trace!("Hello"; "FlashLog"; version = "0.1.0", author = "John Doe");
-        flash_trace!(version = "0.1.0");
-        flash_trace!(version = "0.1.0", author = "John Doe");
-        flash_trace!("topic1"; "message {} {}", 1, 2);
-        flash_trace!("topic2"; "message {} {}", 1, 2; struct_info = 1, struct_info2 = 2);
-        flash_trace!("topic2"; "message {} {}", 1, 2; Hello = Hello::FlashLog);
+        flash_trace_ct!(Hello::World);
+        flash_trace_ct!("Hello");
+        flash_trace_ct!("Hello"; "FlashLog");
+        flash_trace_ct!("Hello"; "FlashLog"; version = "0.1.0");
+        flash_trace_ct!("Hello"; "FlashLog"; version = "0.1.0", author = "John Doe");
+        flash_trace_ct!(version = "0.1.0");
+        flash_trace_ct!(version = "0.1.0", author = "John Doe");
+        flash_trace_ct!("topic1"; "message {} {}", 1, 2);
+        flash_trace_ct!("topic2"; "message {} {}", 1, 2; struct_info = 1, struct_info2 = 2);
+        flash_trace_ct!("topic2"; "message {} {}", 1, 2; Hello = Hello::FlashLog);
         let test_info = TestStruct { test: 1 };
-        flash_trace!("topic2"; "message {} {}", 1, 2; TestStruct = test_info);
+        flash_trace_ct!("topic2"; "message {} {}", 1, 2; TestStruct = test_info);
+        println!("{:?}", test_info);
 
         crate::flush!();